+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 > **LoginRadius Ionic SDK Change Log** provides information regarding what has changed, more specifically what changes, improvements and bug fix has been made to the SDK. For more details please refer to the [LoginRadius API Documention](http://apidocs.loginradius.com/docs/ionic)
 
 
+### Version 3.0.0
+Released on **November 17,  2017**
+
+##### Enhancements
+
+  - Added new VK native login for better native user experience.
+  - Added new risk based authentication API.
+  - Added new email prompt Auto login API.
+  - Added new update security question API.
+  - Added new no registration or password less login, API.
+  - Added new resetpassword by security question API.
+  - Significantly improved code performance.
+
+
 
 ### Version 2.0.1
 Released on **October 03,  2017**